+++ cloud/filestore/libs/storage/fastshard/impl/factory/shard_factory.h
@@ -10,19 +10,6 @@ namespace NCloud::NFileStore::NStorage::NFastShard {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-struct IFileSystemShardFactory
-{
-    virtual ~IFileSystemShardFactory() = default;
-
-    virtual IFileSystemShardPtr CreateShard(
-        const TString& fileSystemId,
-        const NProtoPrivate::TFastShardConfig& config,
-        ui32 shardNo,
-        ui64 generation) = 0;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
 // A mem shard for a mem config, the persistent one for a persistent config,
 // or a stub answering E_NOT_IMPLEMENTED if the runtime is not enabled here.
 IFileSystemShardFactoryPtr CreateFileSystemShardFactory(bool runtimeEnabled);
